Refactored and cleaned up AppellantDetails, renamed to AppellantContactDetails
