+++ steps/identity/appellant-contact-details/content.en.json
@@ -1,1 +1,44 @@
-{}
+{
+  "title": "Enter your contact details",
+  "subtitle": "We’ll use these to send you information about your appeal.",
+  "fields": {
+    "addressLine1": {
+      "title": "Address line 1",
+      "error": {
+        "required": "Please enter Address line 1"
+      }
+    },
+    "addressLine2": {
+      "title": "Address line 2",
+      "error": {
+        "required": "Please enter Address line 2"
+      }
+    },
+    "townCity": {
+      "title": "Town/City",
+      "error": {
+        "required": "Please enter a Town/City"
+      }
+    },
+    "postCode": {
+      "title": "Postcode",
+      "error": {
+        "required": "Please enter a valid postcode"
+      }
+    },
+    "phoneNumber": {
+      "title": "Phone number (optional)",
+      "hint": "We may use this to contact you about your appeal",
+      "error": {
+        "invalid": "Please enter a valid phone number"
+      }
+    },
+    "emailAddress": {
+      "title": "Email address (optional)",
+      "hint": "We’ll use this to email you updates about your appeal",
+      "error": {
+        "invalid": "Please enter a valid email address"
+      }
+    }
+  }
+}
