+++ steps/identity/appellant-contact-details/template.html
@@ -1,1 +1,27 @@
-<h1>APPELLANT CONTACT DETAILS PLACEHOLDER</h1>
+{% extends "look-and-feel/layouts/question.html" %}
+{% from "look-and-feel/components/fields.njk" import textbox, formSection %}
+
+{% block head -%}
+<link href="{{ asset_path }}main.css" media="screen" rel="stylesheet" />
+{% endblock %}
+
+{% set title %}
+    {{ content.title | safe }}
+{% endset %}
+
+{% block fields %}
+
+    {% call formSection() %}
+
+        <p>{{ content.subtitle}}</p>
+
+        {{ textbox(fields.addressLine1,         content.fields.addressLine1.title,          hint="") }}
+        {{ textbox(fields.addressLine2,         content.fields.addressLine2.title,          hint="") }}
+        {{ textbox(fields.townCity,             content.fields.townCity.title,              hint="") }}
+        {{ textbox(fields.postCode,             content.fields.postCode.title,              hint="") }}
+        {{ textbox(fields.phoneNumber,          content.fields.phoneNumber.title,           hint=fields.phoneNumber.hint) }}
+        {{ textbox(fields.emailAddress,         content.fields.emailAddress.title,          hint=fields.emailAddress.hint) }}
+
+    {% endcall %}
+
+{% endblock %}
