--- steps/identity/appellant-details/content.en.json
@@ -1,63 +1,0 @@
-{
-  "titleNoAppointee": "Enter your details",
-  "titleAppointee": "Enter appellants details",
-  "subtitle": "We’ll use these to send you information about your appeal.",
-  "fields": {
-    "firstName": {
-      "title": "First name(s)",
-      "error": {
-        "required": "Please enter your first name"
-      }
-    },
-    "lastName": {
-      "title": "Last name(s)",
-      "error": {
-        "required": "Please enter your last name"
-      }
-    },
-    "niNumber": {
-      "title": "National Insurance number",
-      "error": {
-        "required": "Please enter a valid NI number"
-      }
-    },
-    "addressLine1": {
-      "title": "Address line 1",
-      "error": {
-        "required": "Please enter Address line 1"
-      }
-    },
-    "addressLine2": {
-      "title": "Address line 2",
-      "error": {
-        "required": "Please enter Address line 2"
-      }
-    },
-    "townCity": {
-      "title": "Town/City",
-      "error": {
-        "required": "Please enter a Town/City"
-      }
-    },
-    "postCode": {
-      "title": "Postcode",
-      "error": {
-        "required": "Please enter a valid postcode"
-      }
-    },
-    "appellantPhoneNumber": {
-      "title": "Phone number (optional)",
-      "hint": "We may use this to contact you about your appeal",
-      "error": {
-        "invalid": "Please enter a valid phone number"
-      }
-    },
-    "emailAddress": {
-      "title": "Email address (optional)",
-      "hint": "We’ll use this to email you updates about your appeal",
-      "error": {
-        "invalid": "Please enter a valid email address"
-      }
-    }
-  }
-}
