--- steps/identity/appellant-details/template.html
@@ -1,38 +1,0 @@
-{% extends "look-and-feel/layouts/question.html" %}
-{% from "look-and-feel/components/header.njk" import header %}
-{% from "look-and-feel/components/fields.njk" import textbox, formSection %}
-
-{% block head -%}
-<link href="{{ asset_path }}main.css" media="screen" rel="stylesheet" />
-{% endblock %}
-
-{% set phase="ALPHA" %}
-{% set feedbackLink="https://github.com/hmcts/one-per-page/issues/new" %}
-
-{% set title %}
-    {% if isAppointee %}
-        {{ content.titleAppointee | safe }}
-    {% else %}
-        {{ content.titleNoAppointee | safe }}
-    {% endif %}
-{% endset %}
-
-{% block fields %}
-
-    {% call formSection() %}
-
-        <p>{{ content.subtitle}}</p>
-
-        {{ textbox(fields.firstName,            content.fields.firstName.title,             hint="") }}
-        {{ textbox(fields.lastName,             content.fields.lastName.title,              hint="") }}
-        {{ textbox(fields.niNumber,             content.fields.niNumber.title,              hint="") }}
-        {{ textbox(fields.addressLine1,         content.fields.addressLine1.title,          hint="") }}
-        {{ textbox(fields.addressLine2,         content.fields.addressLine2.title,          hint="") }}
-        {{ textbox(fields.townCity,             content.fields.townCity.title,              hint="") }}
-        {{ textbox(fields.postCode,             content.fields.postCode.title,              hint="") }}
-        {{ textbox(fields.appellantPhoneNumber, content.fields.appellantPhoneNumber.title,  hint=fields.appellantPhoneNumber.hint) }}
-        {{ textbox(fields.emailAddress,         content.fields.emailAddress.title,          hint=fields.emailAddress.hint) }}
-
-    {% endcall %}
-
-{% endblock %}
